Add hexagonal-2d to DynamicSOM

+++ test/PythonBindingTest/DynamicTrainer.cpp
@@ -21,7 +21,7 @@ TEST(GenericPythonBindingTest, DynamicTrainer)
 
     DynamicData data("float32", "cartesian-2d", shape, static_cast<void*>(&ptr[0]));
 
-    std::vector<ssize_t> som_shape{2, 2, 3, 3};
+    std::vector<uint32_t> som_shape{2, 2, 3, 3};
     std::vector<float> som_ptr(16, 0.0f);
 
     DynamicSOM som("float32", "cartesian-2d", "cartesian-2d", som_shape, static_cast<void*>(&som_ptr[0]));
